Make copying links to issues more useful
This now includes the reviewers, which are taken from three sources:

    1. The issue opener
    2. The issue assignees
    3. The person that last marked the API as api-ready-for-review

The site uses the OSPO service to translate GitHub user names to Microsoft
users. If the person isn't a Microsoft employee, it's discarded from the
reviewer set.

When copying links to issue, we're now generating Outlook @mentions which
means sending out invites with the agenda is much simpler -- we no longer
have to manually figure out who to tag and then mention by hand; we just
paste the text and Outlook automatically puts the mentioned folks on the
To: line.